Add custom service describing the Wheel Proxy status
Service UUID: 00001900-7375-7065-7263-7572696f3a29
Characteristics:
 - 00002c00-7375-7065-7263-7572696f3a29:
    Status, Byte, read & notify
      0: Disconnected from wheel, lost connection,
         attempting to reconnect
      1: Active connection
      2: Proxy is stopping
    The characteristic sends notification only after
    requesting so via its descriptor

 - 00002c01-7375-7065-7263-7572696f3a29:
     original device name, String, read

 - 00002c02-7375-7065-7263-7572696f3a29
     original device MAC address, String, read

Note: the "Proxy is stopping" value acts as a workaround to GATT
server not disconnecting clients when stopping as expected.
It follows a user action: either manual disconnection from a wheel
or manual disabling of the Proxy.
It is also useful for the client connected to the Proxy to know that
this device will not come back (as the same advertisment MAC)

In case the app crashes or is updated by the Play Store, the
"Proxy is stopping" value can't/won't be sent.